# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DICOM Gateway API
  description: |
    REST API for DICOM Gateway - A lightweight, HIPAA-compliant Linux DICOM Gateway
    for receiving and forwarding DICOM studies with byte-preserving integrity.
    
    ## Features
    - C-STORE SCP (receiver) and SCU (forwarder) with byte-preserving transmission
    - PostgreSQL-backed job queue with LISTEN/NOTIFY
    - Role-based access control (RBAC)
    - Comprehensive audit logging
    - Prometheus metrics export
    - Configuration management via YAML
  version: 0.1.0
  contact:
    name: DICOM Gateway Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://gateway.example.com/api/v1
    description: Production server
...
paths:
...
  /auth/login:
    post:
      summary: Login
      description: Authenticate user and receive access token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      summary: Get current user
      description: Returns information about the currently authenticated user
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /audit:
    get:
      summary: List audit logs
      description: List audit logs with filtering
      tags:
        - Audit
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
        - name: user_id
          in: query
          schema:
            type: string
        - name: username
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: resource_type
          in: query
          schema:
            type: string
        - name: resource_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /audit/{audit_id}:
    get:
      summary: Get audit log
      description: Get specific audit log entry
      tags:
        - Audit
      security:
        - BearerAuth: []
      parameters:
        - name: audit_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Audit log entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Skip:
      name: skip
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of records to skip
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      description: Maximum number of records to return
...
  schemas:
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        full_name:
          type: string
          nullable: true
        enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
...
    AuditLogResponse:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        ip_address:
          type: string
          nullable: true
        action:
          type: string
        resource_type:
          type: string
          nullable: true
        resource_id:
          type: string
          nullable: true
        status:
          type: string
        error_message:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'